grocerly/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-jazzmin==3.0.1
django-taggit==6.1.0
django-shortuuidfield==0.1.3
pillow==12.0.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
tzdata==2025.2
gunicorn==23.0.0
